+++ Frieght.Api/Infrastructure/FrieghtDbContext.cs
@@ -15,6 +15,7 @@ namespace Frieght.Api.Infrastructure
         public DbSet<Load> Loads => Set<Load>();
         public DbSet<Carrier> Carriers => Set<Carrier>();
         public DbSet<Bid> Bids => Set<Bid>();
+        public DbSet<Shipper> Shipper => Set<Shipper>();
         protected override void OnModelCreating(ModelBuilder modelBuilder)
         {
             modelBuilder.ApplyConfigurationsFromAssembly(Assembly.GetExecutingAssembly());