+++ Changelog.md
@@ -2,4 +2,5 @@
 
 * [Task 43] - Finish Registration and integrate it with the database.
 * [Task 39] - Finish Auto-deployent setup for frieght.api and make it accessible.
-* [Task 39] -  Add shipper info into load entity
+* [Task 40] -  Add shipper info into load entity
+* [Task 41] -  AddLogger and a code to create shipper when the first load is created to avoid violating foreign key violation